update#merge: iter-3 WIP

+++ nodecosmos/src/models/branch/merge/nodes.rs
@@ -5,28 +5,32 @@ use crate::models::description::{find_description, Description, ObjectType};
 use crate::models::node::reorder::ReorderParams;
 use crate::models::node::sort::SortNodes;
 use crate::models::node::{find_update_title_node, Node, PkNode, UpdateTitleNode};
-use crate::models::traits::context::{Context, ModelContext};
+use crate::models::traits::context::ModelContext;
 use crate::models::traits::ref_cloned::RefCloned;
 use crate::models::traits::{Branchable, GroupById, GroupByObjId, Pluck};
 use crate::models::udts::{BranchReorderData, TextChange};
 use charybdis::operations::{DeleteWithCallbacks, Find, InsertWithCallbacks, UpdateWithCallbacks};
-use charybdis::types::{Frozen, List, Uuid};
-use log::{error, warn};
+use charybdis::types::{Frozen, List, Set, Uuid};
+use log::error;
 use scylla::CachingSession;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
+
+enum DelAncNode<'a> {
+    Node(&'a Node),
+    PkNode(&'a PkNode),
+}
 
 #[derive(Serialize, Deserialize)]
 pub struct MergeNodes {
     restored_nodes: Option<Vec<Node>>,
     created_nodes: Option<Vec<Node>>,
+    deleted_nodes: Option<Vec<Node>>,
     reordered_nodes_data: Option<Vec<BranchReorderData>>,
     edited_title_nodes: Option<Vec<UpdateTitleNode>>,
     edited_node_descriptions: Option<Vec<Description>>,
     original_title_nodes: Option<HashMap<Uuid, UpdateTitleNode>>,
     original_nodes_descriptions: Option<HashMap<Uuid, Description>>,
-    title_change_by_object: Option<HashMap<Uuid, TextChange>>,
-    description_change_by_object: Option<HashMap<Uuid, TextChange>>,
 }
 
 impl MergeNodes {
@@ -65,8 +69,23 @@ impl MergeNodes {
         Ok(None)
     }
 
+    pub async fn deleted_nodes(
+        branch: &Branch,
+        db_session: &CachingSession,
+    ) -> Result<Option<Vec<Node>>, NodecosmosError> {
+        if let Some(deleted_node_ids) = &branch.deleted_nodes {
+            let mut deleted_nodes = Node::find_by_ids(db_session, deleted_node_ids).await?;
+
+            deleted_nodes.sort_by_depth();
+
+            return Ok(Some(deleted_nodes));
+        }
+
+        Ok(None)
+    }
+
     pub fn reordered_nodes_data(branch: &Branch) -> Option<List<Frozen<BranchReorderData>>> {
-        if let Some(reordered_nodes) = branch.reordered_nodes.ref_cloned() {
+        if let Some(reordered_nodes) = branch.reordered_nodes.clone() {
             Some(
                 reordered_nodes
                     .into_iter()
@@ -171,6 +190,7 @@ impl MergeNodes {
     pub async fn new(branch: &Branch, data: &RequestData) -> Result<Self, NodecosmosError> {
         let restored_nodes = Self::restored_nodes(&branch, data.db_session()).await?;
         let created_nodes = Self::created_nodes(&branch, data.db_session()).await?;
+        let deleted_nodes = Self::deleted_nodes(&branch, data.db_session()).await?;
         let reordered_nodes_data = Self::reordered_nodes_data(&branch);
         let edited_title_nodes = Self::edited_title_nodes(&branch, data.db_session()).await?;
         let edited_node_descriptions = Self::edited_node_descriptions(&branch, data.db_session()).await?;
@@ -180,13 +200,12 @@ impl MergeNodes {
         Ok(Self {
             restored_nodes,
             created_nodes,
+            deleted_nodes,
             reordered_nodes_data,
             edited_title_nodes,
             edited_node_descriptions,
             original_title_nodes,
             original_nodes_descriptions,
-            title_change_by_object: None,
-            description_change_by_object: None,
         })
     }
 
@@ -225,8 +244,132 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn restore_nodes(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
-        Self::insert_nodes(data, &self.branch, &mut self.restored_nodes).await?;
+    pub async fn check_conflicts(&mut self, data: &RequestData, branch: &Branch) -> Result<(), NodecosmosError> {}
+
+    async fn extract_created_nodes_conflicts(&mut self, db_session: &CachingSession) -> Result<(), NodecosmosError> {
+        if let Some(created_nodes) = &self.created_nodes {
+            for created_node in created_nodes {
+                self.extract_deleted_ancestors(db_session, branch, DelAncNode::Node(created_node))
+                    .await?;
+            }
+        }
+        Ok(())
+    }
+
+    async fn extract_edited_description_nodes_conflicts(
+        &self,
+        db_session: &CachingSession,
+        branch: &Branch,
+    ) -> Result<(), NodecosmosError> {
+        let created_node_ids = branch.created_nodes.ref_cloned();
+        let restored_node_ids = branch.restored_nodes.ref_cloned();
+        let deleted_node_ids = branch.deleted_nodes.ref_cloned();
+        let mut deleted_edited_nodes = HashSet::new();
+
+        if let Some(edit_description_node_ids) = &branch.edited_description_nodes {
+            let edit_description_node_ids = edit_description_node_ids
+                .iter()
+                .filter_map(|id| {
+                    if created_node_ids.contains(id) || deleted_node_ids.contains(id) {
+                        None
+                    } else {
+                        Some(*id)
+                    }
+                })
+                .collect::<Vec<Uuid>>();
+
+            // check ancestors of edited description nodes
+            let desc_branched_nodes =
+                PkNode::find_by_ids_and_branch_id(&db_session, &edit_description_node_ids, branch.id).await?;
+
+            for node in &desc_branched_nodes {
+                self.extract_deleted_ancestors(db_session, DelAncNode::PkNode(&node))
+                    .await?;
+            }
+
+            let original_nodes_ids = PkNode::find_by_ids(&db_session, &edit_description_node_ids)
+                .await?
+                .pluck_id_set();
+
+            let deleted_edited_desc_nodes = edit_description_node_ids
+                .iter()
+                .filter_map(|id| {
+                    if original_nodes_ids.contains(id) || restored_node_ids.contains(id) {
+                        None
+                    } else {
+                        Some(*id)
+                    }
+                })
+                .collect::<Set<Uuid>>();
+
+            deleted_edited_nodes.extend(deleted_edited_desc_nodes);
+        }
+
+        if deleted_edited_nodes.len() > 0 {
+            self.deleted_edited_nodes = Some(deleted_edited_nodes);
+        }
+
+        Ok(())
+    }
+
+    async fn extract_deleted_ancestors(
+        &mut self,
+        db_session: &CachingSession,
+        branch: &mut Branch,
+        del_anc_node: DelAncNode<'_>,
+    ) -> Result<(), NodecosmosError> {
+        let created_node_ids = branch.created_nodes.ref_cloned();
+        let restored_node_ids = branch.restored_nodes.ref_cloned();
+        let deleted_node_ids = branch.deleted_nodes.ref_cloned();
+        let ancestor_id = match del_anc_node {
+            DelAncNode::Node(node) => node.ancestor_ids.ref_cloned(),
+            DelAncNode::PkNode(pk_node) => pk_node.ancestor_ids.ref_cloned(),
+        };
+
+        let branch_ancestor_ids = ancestor_id
+            .iter()
+            .filter_map(|id| {
+                if created_node_ids.contains(id) || restored_node_ids.contains(id) || deleted_node_ids.contains(id) {
+                    None
+                } else {
+                    Some(*id)
+                }
+            })
+            .collect::<Vec<Uuid>>();
+
+        let original_ancestor_ids_set = PkNode::find_by_ids(&db_session, &branch_ancestor_ids)
+            .await?
+            .pluck_id_set();
+
+        let branch_node_deleted_ancestor_ids = branch_ancestor_ids
+            .iter()
+            .filter_map(|id| {
+                if original_ancestor_ids_set.contains(id) {
+                    None
+                } else {
+                    Some(*id)
+                }
+            })
+            .collect::<Set<Uuid>>();
+
+        if branch_node_deleted_ancestor_ids.is_empty() {
+            return Ok(());
+        }
+
+        branch.deleted_ancestors = match self.deleted_ancestors.as_mut() {
+            Some(deleted_ancestors) => {
+                let mut deleted_ancestors = deleted_ancestors.clone();
+                deleted_ancestors.extend(branch_node_deleted_ancestor_ids);
+                Some(deleted_ancestors)
+            }
+            None => Some(branch_node_deleted_ancestor_ids),
+        };
+
+        Ok(())
+    }
+
+    pub async fn restore_nodes(&mut self, data: &RequestData, branch: &Branch) -> Result<(), NodecosmosError> {
+        Self::insert_nodes(data, branch, &mut self.restored_nodes).await?;
 
         Ok(())
     }
@@ -237,8 +380,8 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn create_nodes(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
-        Self::insert_nodes(data, &self.branch, &mut self.created_nodes).await?;
+    pub async fn create_nodes(&mut self, data: &RequestData, branch: &Branch) -> Result<(), NodecosmosError> {
+        Self::insert_nodes(data, branch, &mut self.created_nodes).await?;
 
         Ok(())
     }
@@ -250,32 +393,19 @@ impl MergeNodes {
     }
 
     pub async fn delete_nodes(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
-        let deleted_node_ids = self.branch.deleted_nodes.ref_cloned();
+        if let Some(deleted_nodes) = &mut self.deleted_nodes {
+            let deleted_node_ids = deleted_nodes.clone().pluck_id_set();
 
-        for deleted_node_id in deleted_node_ids.clone() {
-            let deleted_node = Node::find_by_id_and_branch_id(deleted_node_id, deleted_node_id)
-                .execute(data.db_session())
-                .await
-                .ok();
-
-            match deleted_node {
-                Some(mut deleted_node) => {
-                    if let Some(ancestor_ids) = &deleted_node.ancestor_ids {
-                        if ancestor_ids.iter().any(|id| deleted_node_ids.contains(id)) {
-                            // skip deletion of node if it has an ancestor that is also deleted as
-                            // it will be removed in the callback
-                            continue;
-                        }
-
-                        deleted_node.set_merge_context();
-                        deleted_node.delete_cb(data).execute(data.db_session()).await?;
+            for deleted_node in deleted_nodes {
+                if let Some(ancestor_ids) = &deleted_node.ancestor_ids {
+                    if ancestor_ids.iter().any(|id| deleted_node_ids.contains(id)) {
+                        // skip deletion of node if it has an ancestor that is also deleted as
+                        // it will be removed in the callback
+                        continue;
                     }
-                }
-                None => {
-                    warn!(
-                        "Failed to find deleted node with id {} and branch id {}",
-                        deleted_node_id, deleted_node_id
-                    );
+
+                    deleted_node.set_merge_context();
+                    deleted_node.delete_cb(data).execute(data.db_session()).await?;
                 }
             }
         }
@@ -284,25 +414,19 @@ impl MergeNodes {
     }
 
     pub async fn undo_delete_nodes(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
-        let deleted_node_ids = self.branch.deleted_nodes.ref_cloned();
+        if let Some(deleted_nodes) = &mut self.deleted_nodes {
+            let deleted_node_ids = deleted_nodes.clone().pluck_id_set();
 
-        for deleted_node_id in deleted_node_ids.clone() {
-            let deleted_node = Node::find_by_id_and_branch_id(deleted_node_id, deleted_node_id)
-                .execute(data.db_session())
-                .await
-                .ok();
+            for deleted_node in deleted_nodes {
+                if let Some(ancestor_ids) = &deleted_node.ancestor_ids {
+                    if ancestor_ids.iter().any(|id| deleted_node_ids.contains(id)) {
+                        // skip deletion of node if it has an ancestor that is also deleted as
+                        // it will be removed in the callback
+                        continue;
+                    }
 
-            match deleted_node {
-                Some(mut deleted_node) => {
-                    deleted_node.set_original_id();
                     deleted_node.set_merge_context();
                     deleted_node.insert_cb(data).execute(data.db_session()).await?;
-                }
-                None => {
-                    warn!(
-                        "Failed to find deleted node with id {} and branch id {}",
-                        deleted_node_id, deleted_node_id
-                    );
                 }
             }
         }
@@ -381,7 +505,7 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn update_nodes_titles(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
+    pub async fn update_title(&mut self, data: &RequestData, branch: &mut Branch) -> Result<(), NodecosmosError> {
         let original_title_nodes = match self.original_title_nodes.as_ref() {
             Some(nodes) => nodes,
             None => return Ok(()),
@@ -392,7 +516,7 @@ impl MergeNodes {
             None => return Ok(()),
         };
 
-        for mut edited_node_title in edited_node_titles {
+        for edited_node_title in edited_node_titles {
             if let Some(original_node) = original_title_nodes.get(&edited_node_title.id) {
                 if original_node.title != edited_node_title.title {
                     let mut text_change = TextChange::new();
@@ -402,7 +526,8 @@ impl MergeNodes {
                     edited_node_title.set_original_id();
                     edited_node_title.update_cb(data).execute(data.db_session()).await?;
 
-                    self.title_change_by_object
+                    branch
+                        .title_change_by_object
                         .get_or_insert_with(HashMap::default)
                         .insert(edited_node_title.id, text_change);
                 }
@@ -412,7 +537,7 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn undo_update_nodes_titles(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
+    pub async fn undo_update_title(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
         if let Some(original_title_nodes) = &mut self.original_title_nodes {
             for original_title_node in original_title_nodes.values_mut() {
                 original_title_node.update_cb(data).execute(data.db_session()).await?;
@@ -422,7 +547,7 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn update_nodes_description(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
+    pub async fn update_description(&mut self, data: &RequestData, branch: &mut Branch) -> Result<(), NodecosmosError> {
         let edited_node_descriptions = match self.edited_node_descriptions.as_mut() {
             Some(descriptions) => descriptions,
             None => return Ok(()),
@@ -456,7 +581,8 @@ impl MergeNodes {
 
             // update text change with new description
             text_change.assign_new(original.markdown.clone());
-            self.description_change_by_object
+            branch
+                .description_change_by_object
                 .get_or_insert_with(HashMap::default)
                 .insert(object_id, text_change);
         }
@@ -464,7 +590,7 @@ impl MergeNodes {
         Ok(())
     }
 
-    pub async fn undo_update_nodes_description(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
+    pub async fn undo_update_description(&mut self, data: &RequestData) -> Result<(), NodecosmosError> {
         if let Some(original_nodes_description) = &mut self.original_nodes_descriptions {
             for original_node_description in original_nodes_description.values_mut() {
                 // description merge is handled within before_insert callback, so we use update to revert as
